Update compression.py
Some optimizations:
 In the `apply_compressed_weight` function, the code accesses the `compressed_state_dict` multiple times using `compressed_state_dict[full_name]`. It is more efficient to store the value in a variable and use that variable instead.
In the `CLinear` class, the code assigns None to `self.weight` if weight is `None`. This assignment is not necessary because `self.weight` is already None by default.
when calculating the `pad_len`, you can directly use the modulo operator instead of creating a comparison and subtraction. Additionally, the line `data.mul_(scale)` can be replaced with `data *= scale` for better readability.