Update apply_delta.py
In the `apply_delta_low_cpu_mem()` function, the for loop in the if name not in delta_state_dict: clause will iterate over all of the parameters in the base model's state dict, even if the `delta_state_dict` does not contain all of the parameters. This can lead to a significant memory leak, as the `delta_state_dict` will be loaded into memory even if it is not used. I guess we can add a check to the if name not in `delta_state_dict:` clause to see if the parameter is already in the `weight_map` dictionary. If it is, then there is no need to load the parameter from the `delta_state_dict`, as it will already be in the target model's state dict.