Update preprocess_toolllama_data.py
`append_list()` function is called recursively to append the list of instances to the output list. However, each time the function is called, it creates a new list. This means that the output list will contain a nested list of lists, which can be inefficient. A better way to implement the `append_list()` function would be to use a single list like this: